Sync to master

+++ BuildProduct.sh
@@ -6,16 +6,33 @@ SCRIPT=$(readlink -f "$0")
 SCRIPTPATH=$(dirname "$SCRIPT")
 cd "$SCRIPTPATH"
 
-
-VERSION="3.13.0"
+# Utils
+VERSION="3.36.0"
 MAVEN_QUIET=""
-INTELLIJ_VERSION=false
 
-while getopts ":hqv" option; do
+# Eclipse
+BUILD_ECLIPSE=true
+
+# IntelliJ
+BUILD_INTELLIJ=true
+
+INJECT_INTELLIJ_VERSION=false
+
+IJ_VERSION_LATEST=LATEST-EAP-SNAPSHOT
+IJ_DISPLAY_VERSION_LATEST=LATEST-EAP-SNAPSHOT
+IJ_SCALA_VERSION_LATEST=2019.3.7
+
+while getopts "hqve:" option; do
     case $option in
-        h) echo "usage: $0 [-h] [-q] [-v]"; exit ;;
+        h) echo "usage: $0 [-h] [-q] [-v] [-e eclipse/intellij]"; exit ;;
         q) MAVEN_QUIET="-q" ;;
-        v) INTELLIJ_VERSION=true ;;
+        v) INJECT_INTELLIJ_VERSION=true ;;
+        e)
+          shopt -s nocasematch
+          case $OPTARG in 
+            eclipse) BUILD_ECLIPSE=false ;;
+            intellij) BUILD_INTELLIJ=false ;;
+          esac ;;
         ?) echo "error: option -$OPTARG is not implemented"; exit ;;
     esac
 done
@@ -28,7 +45,6 @@ TOSIGNPATH="/c/Signing/ToSign"
 SIGNEDPATH="/c/Signing/Signed"
 ECLIPSE_TOSIGN="/c/jenkins/toSignPackage/eclipse"
 INTELLIJ_TOSIGN="/c/jenkins/toSignPackage/intelliJ"
-CSU_PATH="/c/Signing/CodeSignUtility"
 
 # check dir exists
 if [ ! -d  "$ARTIFACTS_DIR" ]; then
@@ -40,6 +56,7 @@ fi
 set -x
 
 # Build Utils
+echo "Building Utils ..."
 mvn install -f ./Utils/pom.xml -Dmaven.repo.local=./.repository $MAVEN_QUIET
 rm -rf ${TOSIGNPATH}/*
 cp ./Utils/azuretools-core/target/azuretools-core-${VERSION}.jar ${TOSIGNPATH}/
@@ -54,6 +71,7 @@ else
 	exit 1
 fi
 
+# Continue building Utils with signed jars
 cp ${SIGNEDPATH}/azuretools-core-${VERSION}.jar ./Utils/azuretools-core/target/azuretools-core-${VERSION}.jar
 cp ${SIGNEDPATH}/azure-explorer-common-${VERSION}.jar ./Utils/azure-explorer-common/target/azure-explorer-common-${VERSION}.jar
 cp ${SIGNEDPATH}/hdinsight-node-common-${VERSION}.jar ./Utils/hdinsight-node-common/target/hdinsight-node-common-${VERSION}.jar
@@ -64,34 +82,32 @@ mvn install:install-file -Dfile=./Utils/hdinsight-node-common/target/hdinsight-n
 
 mvn install -f ./PluginsAndFeatures/AddLibrary/AzureLibraries/pom.xml -Dmaven.repo.local=./.repository $MAVEN_QUIET
 
-# # Build eclipse plugin
-mvn clean install -f ./PluginsAndFeatures/azure-toolkit-for-eclipse/pom.xml -Dinstrkey=${ECLIPSE_KEY}  $MAVEN_QUIET
-cp ./PluginsAndFeatures/azure-toolkit-for-eclipse/WindowsAzurePlugin4EJ/target/WindowsAzurePlugin4EJ*.zip ./$ARTIFACTS_DIR/WindowsAzurePlugin4EJ.zip
+# Build Eclipse plugin
+if $BUILD_ECLIPSE; then
+  echo "Building Eclipse plugin ..."
+  mvn clean install -f ./PluginsAndFeatures/azure-toolkit-for-eclipse/pom.xml -Dinstrkey=${ECLIPSE_KEY}  $MAVEN_QUIET
+  cp ./PluginsAndFeatures/azure-toolkit-for-eclipse/WindowsAzurePlugin4EJ/target/WindowsAzurePlugin4EJ*.zip ./$ARTIFACTS_DIR/WindowsAzurePlugin4EJ.zip
 
-chmod +x ./tools/IntellijVersionHelper
-
-# Build intellij 2018.1 plugin
-if [ $INTELLIJ_VERSION == "true" ] ; then
-    ./tools/IntellijVersionHelper 2018.1
+   # Extract jars to sign
+  rm -rf ${ECLIPSE_TOSIGN}/*
+  unzip -j artifacts/WindowsAzurePlugin4EJ.zip "**/*.jar" "*.jar" -d ${ECLIPSE_TOSIGN}
 fi
-(cd PluginsAndFeatures/azure-toolkit-for-intellij && ./gradlew clean buildPlugin -s -Papplicationinsights.key=${INTELLIJ_KEY} -Pintellij_version=IC-2018.1 -Pdep_plugins=org.intellij.scala:2018.1.8)
-cp ./PluginsAndFeatures/azure-toolkit-for-intellij/build/distributions/azure-toolkit-for-intellij.zip ./$ARTIFACTS_DIR/azure-toolkit-for-intellij-2018.1.zip
 
-# Build intellij 2018.2 plugin
-if [ $INTELLIJ_VERSION == "true" ] ; then
-    ./tools/IntellijVersionHelper 2018.2
+# Build IntelliJ plugin
+if $BUILD_INTELLIJ; then
+  echo "Building IntelliJ plugin ..."
+  chmod +x ./tools/IntellijVersionHelper
+
+  # Build intellij plugin for latest version
+  if [ $INJECT_INTELLIJ_VERSION == "true" ] ; then
+      ./tools/IntellijVersionHelper $IJ_DISPLAY_VERSION_LATEST
+  fi
+  (cd PluginsAndFeatures/azure-toolkit-for-intellij && ./gradlew clean buildPlugin -s -Papplicationinsights.key=${INTELLIJ_KEY} -Pintellij_version=IC-$IJ_VERSION_LATEST -Pdep_plugins=org.intellij.scala:$IJ_SCALA_VERSION_LATEST)
+  cp ./PluginsAndFeatures/azure-toolkit-for-intellij/build/distributions/azure-toolkit-for-intellij.zip ./$ARTIFACTS_DIR/azure-toolkit-for-intellij-$IJ_DISPLAY_VERSION_LATEST.zip
+
+  # Extract jars to sign
+  rm -rf ${INTELLIJ_TOSIGN}/*
+  unzip -p ./artifacts/azure-toolkit-for-intellij-$IJ_DISPLAY_VERSION_LATEST.zip azure-toolkit-for-intellij/lib/azure-toolkit-for-intellij.jar > ${INTELLIJ_TOSIGN}/azure-toolkit-for-intellij_$IJ_DISPLAY_VERSION_LATEST.jar
 fi
-(cd PluginsAndFeatures/azure-toolkit-for-intellij && ./gradlew clean buildPlugin -s -Papplicationinsights.key=${INTELLIJ_KEY})
-cp ./PluginsAndFeatures/azure-toolkit-for-intellij/build/distributions/azure-toolkit-for-intellij.zip ./$ARTIFACTS_DIR/azure-toolkit-for-intellij-2018.2.zip
-
-# Extract jars to sign
-# intelliJ
-rm -rf ${INTELLIJ_TOSIGN}/*
-unzip -p ./artifacts/azure-toolkit-for-intellij-2018.1.zip azure-toolkit-for-intellij/lib/azure-toolkit-for-intellij.jar > ${INTELLIJ_TOSIGN}/azure-toolkit-for-intellij_2018.1.jar
-unzip -p ./artifacts/azure-toolkit-for-intellij-2018.2.zip azure-toolkit-for-intellij/lib/azure-toolkit-for-intellij.jar > ${INTELLIJ_TOSIGN}/azure-toolkit-for-intellij_2018.2.jar
-
-# Eclipse
-rm -rf ${ECLIPSE_TOSIGN}/*
-unzip -j artifacts/WindowsAzurePlugin4EJ.zip "**/*.jar" "*.jar" -d ${ECLIPSE_TOSIGN}
 
 echo "ALL BUILD SUCCESSFUL"
